requirements.txt
requests==2.31.0
PyGithub==2.1.1
PyNaCl==1.5.0
